requirements.txt
python-telegram-bot
selenium
requests~=2.28.2
PyYAML~=6.0
telegram~=0.0.1
beautifulsoup4~=4.11.2
